+++ src/helpers/fs-ultra.ts
@@ -1,4 +1,5 @@
-import { stat } from "fs/promises";
+import { readdir, stat } from "fs/promises";
+import { join, sep } from "path";
 
 export async function isFolder(path: string) {
 	const res = await stat(path);
@@ -6,6 +7,31 @@ export async function isFolder(path: string) {
 }
 
 export function addTestToFileName(fileName: string) {
-    var dotIndex = fileName.lastIndexOf(".");
+    const dotIndex = fileName.lastIndexOf(".");
     return fileName.substring(0, dotIndex) + ".test" + fileName.substring(dotIndex);
+}
+
+export async function getNearest(fileName: string, from: string): Promise<string | null> {
+	const res = await stat(from);
+	
+	if (res.isDirectory()) {
+		const dir = await readdir(from);
+
+		if (dir.includes(fileName)) {
+			return join(from, fileName);
+		}
+	}
+
+	if (from.split(sep).length === 2) {
+		return null;
+	}
+
+	const index = from.lastIndexOf(sep);
+
+	if (!index) {
+		return null;
+	}
+	
+	const lower = from.slice(0, index);
+	return getNearest(fileName, lower);
 }